setup-environment: check script has been sourced
Check that this script has been invoked correctly (i.e. that it has been
"source"d and not merely run directly). If it has been run directly, don't
exit immediately in case the user specified the "help" option, allow the help
to be displayed, then exit. Adjust how this script is terminated based on
whether it has been sourced or run directly (i.e. use either "return" or
"exit" as required).

This change fixes an infinite loop that is caused if the user runs this script
directly and specifies the help option.

These changes have been tested on: bash, dash, and zsh.

Signed-off-by: Trevor Woerner <trevor.woerner@linaro.org>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>