Revert "setup-environment: check script has been sourced"
This reverts commit a59d9044af96918df5333d1fe64ccb0cc79a40f3.

Some automation scripts, as the ones used in O.S. Systems'
autobuilder, can source the script inside of a control script and the
check using the program name does not work in this case.

This reverts the commit for now, until we find out how to better fix
this issue without breaking external scripts.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>