setup-environment: fix EULA acceptance logic
This commit fixes the handling of the EULA environment variable,
which was being ignored.  This regression was introduced by
41c2b0acbe.

Assumption by this commit: variables are considered set if they are
set according to shell's rules.

So,

  . setup-environment <build-dir>

(assuming EULA is not set in the current environment) or

  EULA= . setup-environment <build-dir>

won't make the FSL EULA be automatically accepted.  However,

  EULA=1 . setup-environment <build-dir>

will, as well as for any value of EULA.

The same rules apply to ACCEPT_FSL_EULA configuration in
local.conf.

Signed-off-by: Mario Domenech Goulart <mario@ossystems.com.br>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>