setup-environment: Keep sources/base files in sync automatically
Currently setup-environment tries to keep the files in sources/base
in sync with the files copied by the manifest. When the files do
get out of sync, the script fails with an error message that many
users will not understand.

We avoid the sync issue in the manifest by linking to the files.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>