setup-environment: Facilitate EULA upgrades for NXP release process
The NXP release process adds a new layer on top of meta-freescale, to
be used for some period of time before the release is integrated into
meta-freescale. The setup for an NXP release is done by an NXP setup
script that wraps the meta-freescale setup script, that in turn
performs the EULA click-through.

Each NXP release does often contain a new EULA license. Until now each
new EULA was handled by having the NXP setup script manually modify
the meta-freescale layer before calling the meta-freescale setup
script. This is a messy solution.

This fix allows the NXP setup script to override the default EULA
file to be used in the click-through.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>